v2.2.1 - Hotfix about file deletion
Delete files that don't exist will throw exceptions, so a judgement is needed.

+++ HappyGenyuanImsactUpdate/Program.cs
@@ -182,7 +182,8 @@ namespace HappyGenyuanImsactUpdate
 
             // Handling with delayed deletions
             foreach (var deletedfile in delete_delays)
-                File.Delete(deletedfile);
+                if (File.Exists(deletedfile))
+                    File.Delete(deletedfile);
 
             //Require Windows 10.0.17763.0+
             if (OperatingSystem.IsWindowsVersionAtLeast(10, 0, 17763, 0))
